+++ Kernel/System/Auth/DB.pm
@@ -150,7 +150,7 @@ sub Auth {
         }
 
         # sha256 pw
-        elsif ( $GetPw =~ m{\A .{64} \z}xms ) {
+        elsif ( $GetPw =~ m{\A [0-9a-f]{64} \z}xmsi ) {
 
             my $SHAObject = Digest::SHA->new('sha256');
 
@@ -195,8 +195,8 @@ sub Auth {
             $Method    = 'bcrypt';
         }
 
-        # fallback: sha1 pw
-        else {
+        # sha1 pw
+        elsif ( $GetPw =~ m{\A [0-9a-f]{40} \z}xmsi ) {
 
             my $SHAObject = Digest::SHA->new('sha1');
 
@@ -206,6 +206,17 @@ sub Auth {
             $SHAObject->add($Pw);
             $CryptedPw = $SHAObject->hexdigest();
             $Method    = 'sha1';
+        }
+
+        # No-13-chars-long crypt pw (e.g. in Fedora28).
+        else {
+            $EncodeObject->EncodeOutput( \$Pw );
+            $EncodeObject->EncodeOutput( \$User );
+
+            # Encode output, needed by crypt() only non utf8 signs.
+            $CryptedPw = crypt( $Pw, $User );
+            $EncodeObject->EncodeInput( \$CryptedPw );
+            $Method = 'crypt';
         }
     }